+++ walletbalance.go
@@ -72,9 +72,10 @@ type WalletBalanceGetResponseBalance struct {
 	// Supported blockchain network names for wallet balance and transaction queries.
 	//
 	// Any of "ethereum", "arbitrum", "avalanche", "base", "tempo", "linea",
-	// "optimism", "polygon", "solana", "zksync_era", "sepolia", "arbitrum_sepolia",
-	// "avalanche_fuji", "base_sepolia", "linea_testnet", "optimism_sepolia",
-	// "polygon_amoy", "solana_devnet", "solana_testnet".
+	// "optimism", "polygon", "bsc", "solana", "tron", "zksync_era", "sepolia",
+	// "arbitrum_sepolia", "avalanche_fuji", "base_sepolia", "linea_testnet",
+	// "optimism_sepolia", "polygon_amoy", "solana_devnet", "solana_testnet",
+	// "tron_nile".
 	Chain            WalletAssetChainNameInput `json:"chain" api:"required"`
 	DisplayValues    map[string]string         `json:"display_values" api:"required"`
 	RawValue         string                    `json:"raw_value" api:"required"`
@@ -105,10 +106,12 @@ const (
 	WalletBalanceGetResponseBalanceAssetEth   WalletBalanceGetResponseBalanceAsset = "eth"
 	WalletBalanceGetResponseBalanceAssetAvax  WalletBalanceGetResponseBalanceAsset = "avax"
 	WalletBalanceGetResponseBalanceAssetPol   WalletBalanceGetResponseBalanceAsset = "pol"
+	WalletBalanceGetResponseBalanceAssetBnb   WalletBalanceGetResponseBalanceAsset = "bnb"
 	WalletBalanceGetResponseBalanceAssetUsdt  WalletBalanceGetResponseBalanceAsset = "usdt"
 	WalletBalanceGetResponseBalanceAssetEurc  WalletBalanceGetResponseBalanceAsset = "eurc"
 	WalletBalanceGetResponseBalanceAssetUsdb  WalletBalanceGetResponseBalanceAsset = "usdb"
 	WalletBalanceGetResponseBalanceAssetSol   WalletBalanceGetResponseBalanceAsset = "sol"
+	WalletBalanceGetResponseBalanceAssetTrx   WalletBalanceGetResponseBalanceAsset = "trx"
 )
 
 type WalletBalanceGetParams struct {
@@ -169,10 +172,12 @@ const (
 	WalletBalanceGetParamsAssetStringEth   WalletBalanceGetParamsAssetString = "eth"
 	WalletBalanceGetParamsAssetStringAvax  WalletBalanceGetParamsAssetString = "avax"
 	WalletBalanceGetParamsAssetStringPol   WalletBalanceGetParamsAssetString = "pol"
+	WalletBalanceGetParamsAssetStringBnb   WalletBalanceGetParamsAssetString = "bnb"
 	WalletBalanceGetParamsAssetStringUsdt  WalletBalanceGetParamsAssetString = "usdt"
 	WalletBalanceGetParamsAssetStringEurc  WalletBalanceGetParamsAssetString = "eurc"
 	WalletBalanceGetParamsAssetStringUsdb  WalletBalanceGetParamsAssetString = "usdb"
 	WalletBalanceGetParamsAssetStringSol   WalletBalanceGetParamsAssetString = "sol"
+	WalletBalanceGetParamsAssetStringTrx   WalletBalanceGetParamsAssetString = "trx"
 )
 
 // Only one field can be non-zero.